# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API Para servir y administrar restaurantes\nhttps://devsstore.net\n[http://swagger.io](http://swagger.io) or on \n[irc.freenode.net, #swagger](http://swagger.io/irc/).\n",
    "version": "1.2.0",
    "title": "MenuXD",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "http://104.248.75.235:5000",
  "basePath": "/api/v1",
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/users/": {
      "get": {
        "description": "Retorna un arreglo con todos los usuarios",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          },
          "400": {
            "description": "Error en la solicitud"
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "No encontrado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "post": {
        "description": "Crea un nuevo usuario",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Creado"
          },
          "400": {
            "description": "Error en la solicitud"
          },
          "401": {
            "description": "No autorizado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "description": "Retorna un usuario por su id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID para obtener un usuario.",
            "required": true,
            "type": "number"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Error en la solicitud"
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "No encontrado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "put": {
        "description": "Actualiza un usuario por su id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID para obtener un usuario.",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Error en la solicitud"
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "No encontrado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "description": "Elimina un usuario por su id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID para obtener un usuario.",
            "required": true,
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Error en la solicitud"
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "No encontrado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/clients/user/{userId}": {
      "get": {
        "description": "Retorna un arreglo con todos los clientes",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID de usuario para todos sus clientes.",
            "required": true,
            "type": "number"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Client"
              }
            }
          },
          "400": {
            "description": "Error en la solicitud"
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "No encontrado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/orders/call/{clientId}/waiter": {
      "get": {
        "description": "Solicita al mozo.",
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "description": "Client id",
            "required": true,
            "type": "number"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Error en la solicitud"
          },
          "401": {
            "description": "No autorizado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/waiters/": {
      "post": {
        "description": "Crea un nuevo mozo",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Waiter"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Creado"
          },
          "400": {
            "description": "Error en la solicitud"
          },
          "401": {
            "description": "No autorizado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/waiters/client/{clientId}": {
      "get": {
        "description": "Retorna un arreglo de mozos por su id",
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "description": "ID para único del cliente.",
            "required": true,
            "type": "number"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Waiter"
              }
            }
          },
          "400": {
            "description": "Error en la solicitud"
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "No encontrado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/waiters/{id}": {
      "get": {
        "description": "Retorna un mozo por su id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID para obtener un mozo.",
            "required": true,
            "type": "number"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Waiter"
            }
          },
          "400": {
            "description": "Error en la solicitud"
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "No encontrado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "put": {
        "description": "Actualiza un mozo por su id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID para actualizar un mozo.",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Waiter"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Error en la solicitud"
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "No encontrado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "description": "Elimina un mozo por su id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID para borrar un mozo.",
            "required": true,
            "type": "number"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Error en la solicitud"
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "No encontrado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "required": [
        "email",
        "role"
      ],
      "properties": {
        "id": {
          "type": "number",
          "example": 2.0
        },
        "email": {
          "type": "string",
          "example": "test@example.com"
        },
        "role": {
          "type": "string",
          "example": "admin"
        },
        "password": {
          "type": "string",
          "example": "Xd123456"
        },
        "image_url": {
          "type": "string",
          "example": "http://example.com/perfil.jpg"
        },
        "confirmed": {
          "type": "boolean",
          "example": true
        },
        "created_at": {
          "type": "string"
        },
        "updated_at": {
          "type": "string"
        }
      }
    },
    "Client": {
      "type": "object",
      "required": [
        "name"
      ],
      "properties": {
        "id": {
          "type": "number",
          "example": 1.0
        },
        "name": {
          "type": "string",
          "example": "Pizza Star"
        },
        "user_id": {
          "type": "number",
          "example": 2.0
        },
        "active": {
          "type": "boolean",
          "example": true
        },
        "picture": {
          "type": "string",
          "example": "http://example.com/picture.png"
        },
        "expire_at": {
          "type": "string"
        },
        "created_at": {
          "type": "string"
        },
        "updated_at": {
          "type": "string"
        }
      }
    },
...
    "Waiter": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 2.0
        },
        "name": {
          "type": "string",
          "example": "max"
        },
        "pin": {
          "type": "string",
          "example": "1234"
        },
        "client_id": {
          "type": "number",
          "example": 2.0
        },
        "created_at": {
          "type": "string"
        },
        "updated_at": {
          "type": "string"
        }
      }
    },
...
  }
}